Synthesizer: Side-issues around refactor out ChildDocuments: Use Document_SideEffect_CreatePatch in there: do that for new documents by default. Call delete validator on Patch. Corrected PatchValidator_Delete; it did not execute the postponed execute. Also in the presentation layer auto-open the document you just created and auto-show DocumentGrid after DocumentClose.

+++ Business/JJ.Business.Synthesizer/PatchManager.cs
@@ -211,6 +211,12 @@ namespace JJ.Business.Synthesizer
         {
             AssertPatchNotNull();
 
+            IValidator validator = new PatchValidator_Delete(Patch, _repositories.PatchRepository);
+            if (!validator.IsValid)
+            {
+                return validator.ToResult();
+            }
+
             Patch.DeleteRelatedEntities(_repositories.OperatorRepository, _repositories.InletRepository, _repositories.OutletRepository, _repositories.EntityPositionRepository);
             Patch.UnlinkRelatedEntities();
             _repositories.PatchRepository.Delete(Patch);
